Insert and read user IDs in URL for link sharing
When clicking on a user in the user select, the user's database ID is added to the URL, and removed when returning to the user list. On initial load with the user ID after /users/, the user with the given ID will be opened immediately. This allows user pages to be referenced by URL for saving or sharing.